Create delphi-code
This Delphi code defines functions for adding, subtracting, and multiplying matrices. Matrices are represented as dynamic arrays of arrays of integers (TMatrix). In the main part of the code, two example matrices A and B are defined, and the operations are performed on them. Finally, the results are printed out.